Things

+++ src/fsized.c
@@ -8,7 +8,6 @@
 #include <p101_c/p101_string.h>
 #include <p101_convert/integer.h>
 #include <p101_fsm/fsm.h>
-#include <p101_posix/p101_string.h>
 #include <p101_posix/p101_unistd.h>
 #include <signal.h>
 #include <stdio.h>
@@ -29,26 +28,25 @@ enum states
     CLEANUP,
 };
 
-static volatile sig_atomic_t exit_flag = 0;    // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
+static volatile sig_atomic_t exit_flag           = 0;    // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
 static volatile sig_atomic_t client_socket_close = 0;    // NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
 
 static void             setup_signal_handlers(void);
 static void             sig_handler(int signal);
 static p101_fsm_state_t parse_arguments(const struct p101_env *env, struct p101_error *err, void *context);
 static p101_fsm_state_t handle_arguments(const struct p101_env *env, struct p101_error *err, void *context);
-static int              socket_create(const struct p101_env *env, struct p101_error *err);
-static void             socket_bind(const struct p101_env *env, struct p101_error *err, int socket_fd, const char *path);
 static p101_fsm_state_t wait_for_request(const struct p101_env *env, struct p101_error *err, void *context);
 static p101_fsm_state_t handle_requests(const struct p101_env *env, struct p101_error *err, void *context);
 static void             handle_new_connection(const struct p101_env *env, struct p101_error *err, int socket_fd, int **client_sockets, nfds_t *max_clients, struct pollfd **fds);
-static void             handle_client_data(const struct p101_env *env, struct p101_error *err, struct pollfd fd, int client_socket);
-static char * read_client_file(const struct p101_env *env, struct p101_error *err, char *filepath);
-static void             respond(const struct p101_env *env, struct p101_error *err, int client_socket, char *message);
+static void             handle_client_data(const struct p101_env *env, struct p101_error *err, int client_socket);
+static char            *read_client_file(const struct p101_env *env, struct p101_error *err, char *filepath);
+static void             respond(const struct p101_env *env, struct p101_error *err, int client_socket, const char *message);
 static void             handle_client_disconnection(const struct p101_env *env, struct p101_error *err, int **client_sockets, nfds_t *max_clients, struct pollfd **fds, nfds_t client_index);
 static p101_fsm_state_t usage(const struct p101_env *env, struct p101_error *err, void *context);
 static p101_fsm_state_t cleanup(const struct p101_env *env, struct p101_error *err, void *context);
 
 #define MSG_LEN 256    // NOLINT(cppcoreguidelines-macro-to-enum, modernize-macro-to-enum)
+#define SIZE_LEN 21    // NOLINT(cppcoreguidelines-macro-to-enum, modernize-macro-to-enum)
 
 static void setup_signal_handlers(void)
 {
@@ -94,9 +92,11 @@ int main(int argc, char *argv[])
         {PARSE_ARGS,       HANDLE_ARGS,      handle_arguments},
         {PARSE_ARGS,       USAGE,            usage           },
         {HANDLE_ARGS,      WAIT_FOR_REQUEST, wait_for_request},
+        {HANDLE_ARGS,      CLEANUP,          cleanup         },
         {WAIT_FOR_REQUEST, HANDLE_REQUESTS,  handle_requests },
         {WAIT_FOR_REQUEST, CLEANUP,          cleanup         },
         {HANDLE_REQUESTS,  WAIT_FOR_REQUEST, wait_for_request},
+        {HANDLE_REQUESTS,  CLEANUP,          cleanup         },
         {USAGE,            CLEANUP,          cleanup         },
         {CLEANUP,          P101_FSM_EXIT,    NULL            }
     };
@@ -154,14 +154,12 @@ int main(int argc, char *argv[])
 
     fsm = p101_fsm_info_create(env, err, "fsized-fsm", fsm_env, fsm_err, NULL);
 
-    // TODO check for error
+    p101_fsm_info_set_did_change_state_notifier(fsm, p101_fsm_info_default_did_change_state_notifier);    // TODO
 
     p101_fsm_run(fsm, &from_state, &to_state, &ctx, transitions, sizeof(transitions));
     p101_fsm_info_destroy(env, &fsm);
 
-    // TODO check for error
-
-    p101_free(fsm_env, fsm_err);
+    free(fsm_env);
 
 free_fsm_error:
     p101_error_reset(fsm_err);
@@ -252,86 +250,58 @@ static p101_fsm_state_t parse_arguments(const struct p101_env *env, struct p101_
 
 p101_fsm_state_t handle_arguments(const struct p101_env *env, struct p101_error *err, void *context)
 {
-    struct contextd *ctx;
-    p101_fsm_state_t next_state;
+    struct contextd   *ctx;
+    p101_fsm_state_t   next_state;
+    struct sockaddr_un addr;
 
     P101_TRACE(env);
     ctx        = (struct contextd *)context;
     next_state = WAIT_FOR_REQUEST;
 
-    ctx->socket_fd = socket_create(env, err);
+    unlink(ctx->arguments->socket_path);
 
-    if(ctx->socket_fd != -1 && p101_error_has_no_error(err))
+    ctx->socket_fd = create_un_socket();
+
+    if(ctx->socket_fd == -1)
     {
-        socket_bind(env, err, ctx->socket_fd, ctx->arguments->socket_path);
-
-        if(p101_error_has_no_error(err))
-        {
-            if(listen(ctx->socket_fd, SOMAXCONN) == -1)
-            {
-                P101_ERROR_RAISE_USER(err, "Failed to listen to socket", ERR_SOCKET);
-            }
-        }
+        P101_ERROR_RAISE_USER(err, "Failed to create to socket", ERR_SOCKET);
+        goto done;
     }
 
-    else if(p101_error_has_error(err))
+    if(init_un_addr(&addr, ctx->arguments->socket_path) == -1)
+    {
+        P101_ERROR_RAISE_USER(err, "Failed to create create unix socket address", ERR_SOCKET);
+        goto done;
+    }
+
+    if(bind(ctx->socket_fd, (struct sockaddr *)&addr, sizeof(addr)) == -1)
+    {
+        P101_ERROR_RAISE_USER(err, "Failed to bind to socket", ERR_SOCKET);
+        goto done;
+    }
+
+    if(listen(ctx->socket_fd, SOMAXCONN) == -1)
+    {
+        P101_ERROR_RAISE_USER(err, "Failed to listen to socket", ERR_SOCKET);
+        goto done;
+    }
+
+    ctx->fds = (struct pollfd *)p101_malloc(env, err, (1) * sizeof(struct pollfd));
+
+    if(p101_error_has_no_error(err))
+    {
+        ctx->fds[0].fd     = ctx->socket_fd;
+        ctx->fds[0].events = POLLIN;
+    }
+
+done:
+
+    if(p101_error_has_error(err))
     {
         next_state = CLEANUP;
     }
 
     return next_state;
-}
-
-static int socket_create(const struct p101_env *env, struct p101_error *err)
-{
-    int socket_fd;
-
-    P101_TRACE(env);
-
-#ifdef SOCK_CLOEXEC
-    socket_fd = socket(AF_UNIX, SOCK_STREAM | SOCK_CLOEXEC, 0);
-#else
-    socket_fd = socket(AF_UNIX, SOCK_STREAM, 0);    // NOLINT(android-cloexec-socket)
-#endif
-
-    if(socket_fd == -1)
-    {
-        P101_ERROR_RAISE_USER(err, "Socket creation failed", ERR_SOCKET);
-    }
-
-    return socket_fd;
-}
-
-static void socket_bind(const struct p101_env *env, struct p101_error *err, int socket_fd, const char *path)
-{
-    struct sockaddr_un addr;
-
-    P101_TRACE(env);
-
-    if(unlink(path) == -1)
-    {
-        P101_ERROR_RAISE_USER(err, "Failed to unlink", ERR_SOCKET);
-    }
-    else
-    {
-        memset(&addr, 0, sizeof(addr));
-        addr.sun_family = AF_UNIX;
-
-        if(strlen(path) >= sizeof(addr.sun_path))
-        {
-            P101_ERROR_RAISE_USER(err, "Socket path too long to accurately bind", ERR_SOCKET);
-        }
-        else
-        {
-            strncpy(addr.sun_path, path, sizeof(addr.sun_path) - 1);
-            addr.sun_path[sizeof(addr.sun_path) - 1] = '\0';
-
-            if(bind(socket_fd, (struct sockaddr *)&addr, sizeof(addr)) == -1)
-            {
-                P101_ERROR_RAISE_USER(err, "Failed to bind socket", ERR_SOCKET);
-            }
-        }
-    }
 }
 
 static p101_fsm_state_t wait_for_request(const struct p101_env *env, struct p101_error *err, void *context)
@@ -368,20 +338,30 @@ static p101_fsm_state_t handle_requests(const struct p101_env *env, struct p101_
     ctx        = (struct contextd *)context;
     next_state = WAIT_FOR_REQUEST;
 
-    handle_new_connection(env, err, ctx->socket_fd, &ctx->client_sockets, &ctx->max_clients, &ctx->fds);
-
-    if(p101_error_has_no_error(err) && ctx->client_sockets != NULL)
+    if(ctx->fds[0].revents & POLLIN)
     {
-        for(nfds_t i = 0; i < ctx->max_clients; i++)
+        handle_new_connection(env, err, ctx->socket_fd, &ctx->client_sockets, &ctx->max_clients, &ctx->fds);
+    }
+
+    if(p101_error_has_error(err))
+    {
+        goto done;
+    }
+
+    for(nfds_t i = 0; i < ctx->max_clients; i++)
+    {
+        if(ctx->client_sockets[i] != -1 && (ctx->fds[i + 1].revents & POLLIN))
         {
-            if(ctx->client_sockets[i] != -1 && (ctx->fds[i + 1].revents & POLLIN))
+            handle_client_data(env, err, ctx->client_sockets[i]);
+
+            if(p101_error_has_no_error(err))
             {
-                handle_client_data(env, err, ctx->fds[i + 1], ctx->client_sockets[i]);
                 handle_client_disconnection(env, err, &ctx->client_sockets, &ctx->max_clients, &ctx->fds, i);
             }
         }
     }
 
+done:
     if(p101_error_has_error(err))
     {
         next_state = CLEANUP;
@@ -392,158 +372,154 @@ static p101_fsm_state_t handle_requests(const struct p101_env *env, struct p101_
 
 void handle_new_connection(const struct p101_env *env, struct p101_error *err, int socket_fd, int **client_sockets, nfds_t *max_clients, struct pollfd **fds)
 {
+    int            new_socket;
+    int           *temp_client_sockets;
+    struct pollfd *new_fds;
+
     P101_TRACE(env);
 
-    if((*fds)[0].revents & POLLIN)
+    new_socket = accept(socket_fd, NULL, NULL);
+
+    if(new_socket == -1)
     {
-        int                new_socket;
-        int               *temp;
-
-        new_socket = accept(socket_fd, NULL, NULL);
-
-        if(new_socket == -1)
-        {
-            P101_ERROR_RAISE_USER(err, "Accept failed", ERR_SOCKET);
-        }
-        else
-        {
-            (*max_clients)++;
-            temp = (int *)p101_realloc(env, err, *client_sockets, sizeof(int) * (*max_clients));
-
-            if(p101_error_has_no_error(err))
-            {
-                struct pollfd *new_fds;
-                *client_sockets                     = temp;
-                (*client_sockets)[*max_clients - 1] = new_socket;
-
-                new_fds = (struct pollfd *)p101_realloc(env, err, *fds, (*max_clients + 1) * sizeof(struct pollfd));
-
-                if(p101_error_has_no_error(err))
-                {
-                    *fds                        = new_fds;
-                    (*fds)[*max_clients].fd     = new_socket;
-                    (*fds)[*max_clients].events = POLLIN;
-                }
-            }
-        }
+        P101_ERROR_RAISE_USER(err, "Accept failed", ERR_SOCKET);
+        goto done;
     }
+
+    (*max_clients)++;
+    temp_client_sockets = (int *)p101_realloc(env, err, *client_sockets, sizeof(int) * (*max_clients));
+
+    if(temp_client_sockets == NULL)
+    {
+        goto done;
+    }
+
+    *client_sockets                     = temp_client_sockets;
+    (*client_sockets)[*max_clients - 1] = new_socket;
+
+    new_fds = (struct pollfd *)p101_realloc(env, err, *fds, (*max_clients + 1) * sizeof(struct pollfd));
+
+    if(new_fds != NULL)
+    {
+        *fds                        = new_fds;
+        (*fds)[*max_clients].fd     = new_socket;
+        (*fds)[*max_clients].events = POLLIN;
+    }
+
+done:
+    return;
 }
 
-void handle_client_data(const struct p101_env *env, struct p101_error *err, struct pollfd fds, int client_socket)
+void handle_client_data(const struct p101_env *env, struct p101_error *err, int client_socket)
 {
+    char    *filename;
     uint16_t filename_length;
+    char    *file_details;
 
     P101_TRACE(env);
+
+    filename     = NULL;
+    file_details = NULL;
 
     if(safe_read(client_socket, &filename_length, sizeof(uint16_t), true) < 0)
     {
         respond(env, err, client_socket, "Failed to parse size of filename");
+        goto done;
     }
-    else
+
+    filename = (char *)p101_malloc(env, err, filename_length);
+
+    if(p101_error_has_error(err))
     {
-        char *filename;
-        filename = (char *)p101_malloc(env, err, filename_length);
-
-        if(p101_error_has_no_error(err))
-        {
-            if(safe_read(client_socket, filename, filename_length, true) < 0)
-            {
-                respond(env, err, client_socket, "Failed to parse filename");
-            }
-            else
-            {
-                char *msg;
-                msg = read_client_file(env, err, filename);
-
-                if(p101_error_has_no_error(err) && msg != NULL)
-                {
-                    respond(env, err, client_socket, filename);
-                }
-
-                p101_free(env, msg);
-            }
-        }
-
-        p101_free(env, filename);
+        goto done;
     }
+
+    if(safe_read(client_socket, filename, filename_length, true) < 0)
+    {
+        respond(env, err, client_socket, "Failed to parse filename");
+        goto done;
+    }
+
+    file_details = read_client_file(env, err, filename);
+
+    if(p101_error_has_no_error(err) && file_details != NULL)
+    {
+        respond(env, err, client_socket, filename);
+    }
+
+done:
+    p101_free(env, file_details);
+    p101_free(env, filename);
 }
 
-static char * read_client_file(const struct p101_env *env, struct p101_error *err, char *filepath)
+static char *read_client_file(const struct p101_env *env, struct p101_error *err, char *filepath)
 {
-    int client_fd;
-    char *response;
+    int         client_fd;
+    char       *response;
+    struct stat file_stat;
+    char       *tmp_response;
 
     P101_TRACE(env);
 
-    response = NULL;
-    client_fd = open(filepath, O_RDONLY);
+    response  = NULL;
+    client_fd = open(filepath, O_RDONLY | O_CLOEXEC);
 
     if(client_fd == -1)
     {
         response = concat_string("Could not open file: ", filepath);
-    }
-    else
-    {
-        struct stat file_stat;
-
-        if(fstat(client_fd, &file_stat) == -1)
-        {
-            response = concat_string("Could not create fstat of: ", filepath);
-        }
-        else
-        {
-            char *tmp_response;
-            tmp_response = concat_string(filepath, " Size in bytes: ");
-            if(tmp_response != NULL)
-            {
-                char size[sizeof(file_stat.st_size)];
-                snprintf(size, sizeof(size), "%lu", file_stat.st_size);
-                response = concat_string(tmp_response, size);
-            }
-        }
-
+        goto done;
     }
 
-    if(p101_close(env, err, client_fd) == -1)
+    if(fstat(client_fd, &file_stat) == -1)
     {
-        P101_ERROR_RAISE_USER(err, "Failed to client file", ERR_OTHER);
+        response = concat_string("Could not create fstat of: ", filepath);
+        goto done;
     }
-    client_fd = -1;
 
-    if (response == NULL)
+    tmp_response = concat_string(filepath, " Size in bytes: ");
+
+    if(tmp_response != NULL)
     {
-        P101_ERROR_RAISE_ERRNO(err, errno);
+        char size[SIZE_LEN];
+        snprintf(size, sizeof(size), "%ld", file_stat.st_size);
+        response = concat_string(tmp_response, size);
+    }
+
+done:
+    if(client_fd > 0 && p101_close(env, err, client_fd) == -1)
+    {
+        p101_error_reset(err);
+        P101_ERROR_RAISE_USER(err, "Failed to close file client sent", ERR_OTHER);
+    }
+    if(response == NULL && p101_error_has_no_error(err))
+    {
+        P101_ERROR_RAISE_USER(err, "Failed to allocate memory for message to send to client", ERR_OTHER);
     }
 
     return response;
 }
 
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wunused-parameter"
-
-static void respond(const struct p101_env *env, struct p101_error *err, int client_socket, char *message)
+static void respond(const struct p101_env *env, struct p101_error *err, int client_socket, const char *message)
 {
+    uint16_t    message_length;
+    const char *actual_message;
+
     P101_TRACE(env);
 
-    if(strlen(message) > UINT16_MAX)
+    if(strlen(message) + 1 > UINT16_MAX)
     {
-        const char too_big_msg[] = "Response too big to send";
-        const uint16_t too_big_msg_len = strlen(too_big_msg);
-
+        actual_message = "Response too big to send";
         printf("Response \"%s\" too large to send.\n", message);
-
-        safe_write(client_socket, &too_big_msg_len, sizeof(uint16_t));
-        safe_write(client_socket, too_big_msg, sizeof(uint16_t));
     }
     else
     {
-        uint16_t safe_message_length;
-
-        safe_message_length = (uint16_t)strlen(message);
-
-        safe_write(client_socket, &safe_message_length, sizeof(uint16_t));
-        safe_write(client_socket, message, safe_message_length);
+        actual_message = message;
     }
+
+    message_length = (uint16_t)strlen(message) + 1;
+
+    safe_write(client_socket, &message_length, sizeof(uint16_t));
+    safe_write(client_socket, actual_message, message_length);
 
     if(client_socket_close == 1)
     {
@@ -554,10 +530,12 @@ static void respond(const struct p101_env *env, struct p101_error *err, int clie
     {
         printf("Successfully sent response \"%s\" to client.\n", message);
     }
+
+    if(shutdown(client_socket, SHUT_RD | SHUT_WR))
+    {
+        P101_ERROR_RAISE_USER(err, "Failed shutdown client socket", ERR_SOCKET);
+    }
 }
-
-#pragma GCC diagnostic pop
-
 
 static void handle_client_disconnection(const struct p101_env *env, struct p101_error *err, int **client_sockets, nfds_t *max_clients, struct pollfd **fds, nfds_t client_index)
 {
@@ -567,8 +545,9 @@ static void handle_client_disconnection(const struct p101_env *env, struct p101_
 
     disconnected_socket = (*client_sockets)[client_index];
 
-    if(p101_close(env, err, disconnected_socket) == -1)
+    if(disconnected_socket > 0 && p101_close(env, err, disconnected_socket) == -1)
     {
+        p101_error_reset(err);
         P101_ERROR_RAISE_USER(err, "Failed to close client connection", ERR_SOCKET);
     }
     else
@@ -611,7 +590,8 @@ static p101_fsm_state_t usage(const struct p101_env *env, struct p101_error *err
 
     fprintf(stderr, "Usage: %s [-h] <socket-path> \n", ctx->arguments->program_name);
     fputs("Options:\n", stderr);
-    fputs(" -h Display this help message\n", stderr);
+    fputs("  -h                Display this help message and exit\n", stderr);
+    fputs("  <socket_path>     Path to the domain socket (required)\n", stderr);
 
     return CLEANUP;
 }
@@ -640,7 +620,6 @@ static p101_fsm_state_t cleanup(const struct p101_env *env, struct p101_error *e
 
     free(ctx->fds);
 
-    // Cleanup and close all client sockets
     for(size_t i = 0; i < ctx->max_clients; i++)
     {
         if(ctx->client_sockets[i] > 0)
@@ -661,7 +640,7 @@ static p101_fsm_state_t cleanup(const struct p101_env *env, struct p101_error *e
 
     free(ctx->client_sockets);
 
-    if(p101_close(env, err, ctx->socket_fd) == -1)
+    if(ctx->socket_fd > 0 && p101_close(env, err, ctx->socket_fd) == -1)
     {
         fputs(p101_error_get_message(err), stderr);
         fputc('\n', stderr);
@@ -672,7 +651,10 @@ static p101_fsm_state_t cleanup(const struct p101_env *env, struct p101_error *e
         ctx->socket_fd = -1;
     }
 
-    unlink(ctx->arguments->socket_path);
+    if(ctx->arguments->socket_path != NULL)
+    {
+        unlink(ctx->arguments->socket_path);
+    }
 
     return P101_FSM_EXIT;
 }
